+++ internal/mktorrent/mktorrent.go
@@ -16,7 +16,6 @@ import (
 )
 
 var DefaultAnnounce = [][]string{
-	{"udp://t.256.fi:6969/announce"},
 	{"https://parrot.lnx.fi/announce"},
 	{"udp://tracker.pirateface.co:6969/announce"},
 	{"udp://tracker.opentrackr.org:1337/announce"},